+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2025-04-30
+
 ### Added
+
+- Added support for label management in the `cockroach_cluster` and
+  `cockroach_folder` resources.
 
 - Added `full_version` attribute to the cluster data source and resource for
   fetching the full version string. (e.g. v25.1.0)